backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23

# Database
psycopg2-binary==2.9.9

# Authentication (specific versions)
python-jose[cryptography]
passlib[bcrypt]

# Configuration
python-dotenv==1.0.0

# Utilities
python-multipart
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0